4. Тесты для стороны прямоугольника. [#159460]

+++ test/ru/job4j/condition/SqAreaTest.java
@@ -7,16 +7,16 @@ public class SqAreaTest {
 
     @Test
     public void whenP6K2Square2() {
-        int p = 6;
+        double p = 6;
         double k = 2;
-        int expected = 2;
+        double expected = 2;
         double out = SqArea.square(p, k);
         Assert.assertEquals(expected, out, 0.01);
     }
 
     @Test
     public void whenP10K7Square273() {
-        int p = 10;
+        double p = 10;
         double k = 7;
         double expected = 2.73;
         double out = SqArea.square(p, k);
@@ -25,7 +25,7 @@ public class SqAreaTest {
 
     @Test
     public void whenP3K11Square017() {
-        int p = 3;
+        double p = 3;
         double k = 11;
         double expected = 0.17;
         double out = SqArea.square(p, k);
